Issue #430. Extracted class RawInmateData to own file
